Address PR comments: Move `FavIconImageRenderer` into it's own file
